edit button and cancel edit button functional without error codes.
