+++ src/scripts/Events/EventsForm.js
@@ -83,47 +83,3 @@ export const renderEventForm = () => {
     `
 };
 
-let contentEditTarget;
-
-export const editEventForm = (eventObj) => {
-    contentEditTarget.innerHTML = `
-    <div class="edit-form" id="edited-event>
-    <section class="event-form">
-    <h3>Edit Event</h3>
-        <fieldset>
-        <label>Title</label>
-        <input type="text"
-        id="event-title" value="${eventObj.title}"></input>
-        </fieldset>
-
-        <fieldset>
-        <label>Description</label>
-        <input type="text"
-        id="event-description" value="${eventObj.description}"></input>
-        </fieldset>
-
-        <fieldset>
-        <label>Location</label>
-        <input type="text"
-        id="event-location" value="${eventObj.location}"></input>
-        </fieldset>
-
-        <fieldset>
-        <label>Time</label>
-        <input type="time"
-        id="event-time" value="${eventObj.time}"></input>
-        </fieldset>
-
-        <fieldset>
-        <label>Date</label>
-        <input type="date"
-        id="event-date" value="${eventObj.date}"></input>
-        </fieldset>
-    
-        <br>
-        <button type="button" id="saveEvent--${eventObj.id}" value="save">Save Event</button>
-        <button type="button" id="editCancel">Cancel Edit</button>
-        <div class="current-event"></div>
-    </div>
-    `
-};
